Refactor planet data catalog
